feat: Cut main file in three files
